Move Swapper and StdInSwapper
